build: migrate from Nuitka onefile to pyproject.toml + python-installer
Nuitka froze a pip-installed PyGObject into the binary; the frozen
gi/overrides asserted against the live system GLib and crashed when
glib2 evolved (#13, #14). Wheel + python-installer keeps gi in lockstep
with the user's system python-gobject by construction.

- Add pyproject.toml with [project.scripts] entry point
- Extract main() into clipse_gui/cli.py so the entry point is importable
- Reduce root clipse-gui.py to a thin dev shim
- Drop Nuitka and the PyGObject pin from requirements.txt
