+++ pyproject.toml
@@ -1,3 +1,47 @@
+[build-system]
+requires = ["setuptools>=64", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "clipse-gui"
+description = "A GTK3 GUI for the clipse clipboard manager"
+readme = "README.md"
+requires-python = ">=3.10"
+license = { file = "LICENSE" }
+authors = [{ name = "D7OMDEV", email = "hello@d7om.dev" }]
+keywords = ["clipboard", "clipse", "gtk", "gtk3", "linux"]
+classifiers = [
+    "Development Status :: 4 - Beta",
+    "Environment :: X11 Applications :: GTK",
+    "Intended Audience :: End Users/Desktop",
+    "License :: OSI Approved :: MIT License",
+    "Operating System :: POSIX :: Linux",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3 :: Only",
+    "Topic :: Desktop Environment",
+    "Topic :: Utilities",
+]
+# Runtime deps are intentionally empty: PyGObject/GTK ship via the system
+# package manager (python-gobject + gtk3) so they stay in lockstep with the
+# system's GLib. Bundling them is what caused #13 / #14.
+dependencies = []
+dynamic = ["version"]
+
+[project.urls]
+Homepage = "https://github.com/d7omdev/clipse-gui"
+Repository = "https://github.com/d7omdev/clipse-gui"
+Issues = "https://github.com/d7omdev/clipse-gui/issues"
+
+[project.scripts]
+clipse-gui = "clipse_gui.cli:main"
+
+[tool.setuptools.dynamic]
+version = { attr = "clipse_gui.__version__" }
+
+[tool.setuptools.packages.find]
+include = ["clipse_gui*"]
+exclude = ["tests*", "aur*", "docs*"]
+
 [tool.pyright]
 reportMissingModuleSource = false
 reportOptionalMemberAccess = true
